Add formality/certainty/urgency/enthusiasm traits and qwen2.5 0.5B/1.5B-instruct/3B/7B models
Extends the trait and model registries for the scale-axis + multi-trait sweep, adds
a focal-layer-only storage policy (with an fp32 covariance fix for the 7B model),
and adds run_analyses.py / collect_summary.py to orchestrate and summarize the sweep.

+++ src/lib/representations.py
@@ -222,6 +222,8 @@ def extract_representations(
     cov_path = cov_path or out_dir / "unembeddings_covariance.pt"
     cov_path.parent.mkdir(parents=True, exist_ok=True)
     if not cov_path.exists():
-        torch.save(unembedding_covariance(model), cov_path)
+        # Saved as float32: at 7B-model hidden sizes (D~3584) float64 would exceed
+        # GitHub's 100MB file limit. Computed in float64 above for precision.
+        torch.save(unembedding_covariance(model).to(torch.float32), cov_path)
         print(f"Saved {cov_path}")
     return out_dir
